+++ modules/common/src/utils/getHtmlInput.ts
@@ -1,0 +1,6 @@
+const getHtmlInput = (
+  ref: React.RefObject<HTMLElement | undefined>
+): HTMLInputElement | null | undefined =>
+  ref?.current?.getElementsByTagName('input').item(0);
+
+export default getHtmlInput;
